+++ services/web/server/src/simcore_service_webserver/rest/rest_plugin.py
@@ -16,13 +16,13 @@ from servicelib.aiohttp.rest_middlewares import (
     error_middleware_factory,
 )
 
+from .._constants import APP_OPENAPI_SPECS_KEY, APP_SETTINGS_KEY
+from .._meta import API_VTAG, api_version_prefix
+from ..security.plugin import setup_security
 from . import rest_handlers
-from ._constants import APP_OPENAPI_SPECS_KEY, APP_SETTINGS_KEY
-from ._meta import API_VTAG, api_version_prefix
 from .rest_healthcheck import HealthCheck
 from .rest_settings import RestSettings, get_plugin_settings
 from .rest_utils import get_openapi_specs_path, load_openapi_specs
-from .security.plugin import setup_security
 
 log = logging.getLogger(__name__)
