+++ services/web/server/src/simcore_service_webserver/rest/rest_schemas_base.py
