+++ services/web/server/src/simcore_service_webserver/rest/rest_settings.py
@@ -1,7 +1,7 @@
 from aiohttp import web
 from settings_library.base import BaseCustomSettings
 
-from ._constants import APP_SETTINGS_KEY
+from .._constants import APP_SETTINGS_KEY
 
 
 class RestSettings(BaseCustomSettings):
