+++ services/web/server/tests/unit/conftest.py
@@ -20,7 +20,7 @@ from models_library.projects import Project
 from openapi_core.schema.specs.models import Spec as OpenApiSpecs
 from pytest_simcore.helpers.utils_dict import ConfigDict
 from pytest_simcore.helpers.utils_projects import empty_project_data
-from simcore_service_webserver.rest_utils import (
+from simcore_service_webserver.rest.rest_utils import (
     get_openapi_specs_path,
     load_openapi_specs,
 )
