+++ services/web/server/tests/unit/isolated/test_rest.py
@@ -14,7 +14,7 @@ from pytest_simcore.helpers.typing_env import EnvVarsDict
 from pytest_simcore.helpers.utils_assert import assert_status
 from servicelib.aiohttp.application import create_safe_application
 from simcore_service_webserver.application_settings import setup_settings
-from simcore_service_webserver.rest import setup_rest
+from simcore_service_webserver.rest.rest_plugin import setup_rest
 from simcore_service_webserver.security.plugin import setup_security
